+++ Engine/Extended/NativeInternal.h
@@ -10,6 +10,8 @@ size_t ME_WritePresentation(void *out, size_t capacity);
 
 size_t ME_WriteMaterials(void *out, size_t capacity);
 size_t ME_WriteBlendTables(void *out, size_t capacity);
+void ME_InitBlendTables(void);
+unsigned ME_BlendTableIndex(const unsigned char *table);
 
 void ME_AudioEnable(void);
 void ME_AudioTick(void);